requirements.txt
django==1.11.3
pillow
mysqlclient
djangorestframework==3.8.0
markdown
django-filter
django-reversion
django-formtools
future
httplib2
six
requests
coreapi
django-cors-headers
xlwt
xlsxwriter
drf-extensions
django-redis
pycryptodome
raven
djangorestframework-jwt
django-crispy-forms